Add lttng-ctl SWIG python bindings
The lttng-tools Python module can be used to directly control the
lttng-tools API inside Python, using 'import lttng'.

Therefore, it becomes possible to create a trace, add events, start/stop
tracing, destroy a session and so on from within Python.

SWIG >= 2.0 is used to create the wrapper and its 'warning md variable
unused' bug is patched in Makefile.am.

In the interface file, struct and enum are directly copied from lttng.h
(all changes must be made in both files).

To install with python bingings, configure using
    --enable-python-bindings

Please note that this extra feature is added in extras/ hence not
supported and not yet shipped by default with the tarball or packages.

Signed-off-by: Danny Serres <danny.serres@efficios.com>
Signed-off-by: Yannick Brosseau <yannick.brosseau@gmail.com>
Signed-off-by: David Goulet <dgoulet@efficios.com>