Fix: remove adjustment on written var for metadata
This was generating the following error each time metadata was streamed
over the network.

    Error: Error writing to tracefile

The return value was off by 8 bytes which is the header stream id size.
However, the "written" variable is not used in the metadata command to
the relayd so there is no use of updating it to fit the expected len of
the mmap() caller.

Signed-off-by: David Goulet <dgoulet@efficios.com>