New URI switching tests
test_uri_switch_localhost_folder:
Random switching of the ouput folder for streaming on localhost.

test_uri_switch_file_network:
Create a session with a file:// URI and enable the consumer with a
net:// URI.

test_uri_switch_network_file:
Create a session with a net:// URI and enable the consumer with a
file:// URI.

All tests support IPv4 and IPv6 streaming targets.

Signed-off-by: Christian Babeux <christian.babeux@efficios.com>
Signed-off-by: David Goulet <dgoulet@efficios.com>