Fix: UST app major version check
Makes a direct match between the major version of the application and
the one supported by lttng-tools (==).

Note that this version numbering is an internal scheme between UST and
lttng-tools. It has no link whatsoever with the public version of the
toolchain.

The 2.1.0-rc1 release commit will make sure this version corresponds to
the UST rc1.

CC: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Signed-off-by: David Goulet <dgoulet@efficios.com>