Move code out of main.c and fix some issues
This is a "code location refactoring". The file main.c was getting a bit
too HUGE so this commit splits the command subsystem of the session
daemon into cmd.c/.h and all helper functions used.

This commit also fixes missing rcu read side lock and the consumer
output hashtable memory leak on destroy.

Minor changes were made to functions call to pass needed variables from
the main.c to the command subsystem.

The behavior was not changed nor altered in any way.

Signed-off-by: David Goulet <dgoulet@efficios.com>