Fix: remove set subdir call that uses bad ptr
When copying a consumer object from the tracing session to a specific
domain session, the original consumer of the session was modified by
appending the session name and datetime once again.

We ended up removing the call since at this stage in the code, the
session name and date is already appended if necessary.

Reported-by: <bernd.hufmann@ericsson.com>
Signed-off-by: David Goulet <dgoulet@efficios.com>