Set hidden visibility for calls used in lttng-ctl
The liblttng-ctl had multiple dynamic symbols that are only used inside
the lttng-tools code tree and should not be exposed to the user.

This commit sets every function calls found with "objdump -T" with a
hidden visibility attribute.

The strpbrk_or_eos() is set static and the struct net_families as well
so we don't export those two symbols.

At this stage there is still the filter_* calls that need to be fixed.

Signed-off-by: David Goulet <dgoulet@efficios.com>