Fix filter: fix recent regressions
Fix:
commit d93c4f1ffcffa73102e3299276f2f83951a68c36
Author: Christian Babeux <christian.babeux@efficios.com>
Date:   Thu Sep 6 13:40:19 2012 -0400

    Fix: Accept bytecode of length 65536 bytes

Broke the filter: it changed the reloc table format, without knowledge
of UST, and without good reason: it should stay { uint16_t, string } --
it does not need a uint32_t offset. Add a check for the value returned
by bytecode_get_len() when populating the reloc table.

Fix:
commit 5ddb0a08c5c3ca917b025032b864d78e53c7c68a
Author: Christian Babeux <christian.babeux@efficios.com>
Date:   Mon Aug 27 14:48:21 2012 -0400

    Fix: Generation of bytecode longer than 32768 bytes fails

Mixed the concepts of "len" and "alloc_len". It was therefore not
checking the bounds correctly, and was not zeroing the memory range
expected to be zeroed, causing out-of-bound array access.

Acked-by: Christian Babeux <christian.babeux@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Signed-off-by: David Goulet <dgoulet@efficios.com>