Fix: Lib lttng-ctl on error returns lttng code
The library was returning negative value that were NOT a lttng error
code documented in lttng-error.h. This was problematic because the
lttng_strerror() could not recognized the returned code and thus the
caller was clueless about the error.

This patch makes sure every lttng-ctl calls return either the correct
documented positive value or a negative lttng error code on error that
can be directly translate by lttng_strerror() which takes a negative
lttng code.

Small fix: enable event command, when hitting a loglevel error now
returns a lttng error code and does not print the help anymore.

Fixes #337

Signed-off-by: David Goulet <dgoulet@efficios.com>